+++ mne/montages/__init__.py
@@ -1,0 +1,3 @@
+"""Module to handle EEG montages"""
+
+from .montage import read_montage, apply_montage